requirements.txt
pyVCF==0.6.8
requests==2.25.1
